# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Taxly API",
        "description": "API documentation for Taxly platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://taxly.test",
            "description": "Local Server"
        },
        {
            "url": "https://taxly.ng",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new tenant, organization, and user",
                "operationId": "fb835c97347e657e2a627790c4c5817a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "tenant_name": {
                                        "type": "string",
                                        "example": "Acme Corp"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "your firs e-invoice login email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "your firs e-invoice login password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "your firs e-invoice login password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tenant, organization, and user created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tenant, organization, and user created successfully",
                                    "success": true,
                                    "data": {
                                        "tenant": {
                                            "id": 1,
                                            "name": "Acme Corp",
                                            "email": "john@example.com",
                                            "brand": "Acme Brand",
                                            "domain": "acme.com",
                                            "entity_id": "123456"
                                        },
                                        "organization": {
                                            "id": 1,
                                            "service_id": "svc-001",
                                            "registration_number": "REG-123456",
                                            "email": "john@example.com",
                                            "phone": "+1234567890",
                                            "street_name": "123 Main St",
                                            "city_name": "Metropolis",
                                            "postal_zone": "12345",
                                            "description": "Main office"
                                        },
                                        "user": {
                                            "id": 1,
                                            "name": "John Doe",
                                            "email": "john@example.com"
                                        },
                                        "token": "1|XyzExampleToken..."
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "fb66b153966d9fac2e999136ef0ade67",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "your firs e-invoice login email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "your firs e-invoice login password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "User logged in successfully",
                                    "success": true,
                                    "data": {
                                        "user": {
                                            "id": 1,
                                            "name": "John Doe",
                                            "email": "john@example.com"
                                        },
                                        "token": "1|XyzExampleToken..."
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get current authenticated user",
                "operationId": "5389aaa95512ef0b6ec88f262ecee7f0",
                "responses": {
                    "200": {
                        "description": "User details",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "User retrieved successfully",
                                    "success": true,
                                    "data": {
                                        "id": 1,
                                        "name": "John Doe",
                                        "email": "john@example.com",
                                        "organization": {
                                            "id": 1,
                                            "legal_name": "Head Office",
                                            "tenant": {
                                                "id": 1,
                                                "name": "Acme Corp"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "307ed0512855922b2280792376bb2095",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the token from login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}